requirements.txt
streamlit
pydantic
google-genai
ollama
groq
python-dotenv
rich
